[Audio][Data] Float32 instead of float64 for melgrams; small things
Cut disk space in half. Also add more fields to tfrecords such as
energy and key.

Also change path for cluster data in views

Test plan:
1) Once done copying tfrecords to tfrecords_backup, run modify records with
tfrecords_backup hardcoded into function
- python tasks/audio_sent/prepare_data.py --modify_tfrecords
2) Check size difference (du -sh). Should be ~half of 224 GB
3) Run mean and std, BUT FIRST, change 'tfrecords' in dirpath to 'tfrecords_backup'
- Also have to change np.fromstring(dtype=np.float32)
- python tasks/audio_sent/prepare_data.py --precompute_numpts_and_meanstd_from_tfrecords
4) Then, modify datasets.py to a) use tfrecords_backup path, and b) decode_raw using tf.float32
5) Then train on Shannon to make sure that it still runs
- python tasks/audio_sent/main.py -bs 32 -lr 0.001 --optim adam -e 100 --gpus 2,3
6) Once we're happy it's running, remove tfrecords/, rename tfrecords_backup as tfrecords
7) Change tfrecords_backup back to tfrecords in precompute_numpts_and_meanstd_from_tfrecords, etc.
8) Run python tasks/audio_sent/prepare_data.py --write_spotify_to_tfrecords

+++ emotion/tasks/audio_sent/datasets.py
@@ -141,7 +141,8 @@ class SpotifyDataset(Dataset):
                 'valence_reg': tf.FixedLenFeature([], tf.float32)
             })
 
-        feats = tf.decode_raw(features['log_melgram'], tf.float64)
+        # feats = tf.decode_raw(features['log_melgram'], tf.float64)
+        feats = tf.decode_raw(features['log_melgram'], tf.float32)
         feats = tf.cast(feats, tf.float32)
         # feats = tf.decode_raw(features['log_melgram'], tf.uint8)
         feats = tf.reshape(feats, MELGRAM_30S_SIZE)
